+++ src/ChunkReader.php
@@ -2,15 +2,18 @@
 
 namespace Maatwebsite\Excel;
 
+use Illuminate\Bus\Queueable;
+use Illuminate\Container\Container;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\PendingDispatch;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\Jobs\SyncJob;
 use Illuminate\Support\Collection;
 use Maatwebsite\Excel\Concerns\ShouldQueueWithoutChain;
 use Maatwebsite\Excel\Concerns\WithChunkReading;
 use Maatwebsite\Excel\Concerns\WithEvents;
 use Maatwebsite\Excel\Concerns\WithLimit;
 use Maatwebsite\Excel\Concerns\WithProgressBar;
-use Maatwebsite\Excel\Events\BeforeImport;
 use Maatwebsite\Excel\Files\TemporaryFile;
 use Maatwebsite\Excel\Imports\HeadingRowExtractor;
 use Maatwebsite\Excel\Jobs\AfterImportJob;
@@ -21,14 +24,24 @@ use Throwable;
 class ChunkReader
 {
     /**
+     * @var Container
+     */
+    protected $container;
+
+    public function __construct(Container $container)
+    {
+        $this->container = $container;
+    }
+
+    /**
      * @param  WithChunkReading  $import
      * @param  Reader  $reader
      * @param  TemporaryFile  $temporaryFile
-     * @return \Illuminate\Foundation\Bus\PendingDispatch|null
+     * @return PendingDispatch|Collection|null
      */
     public function read(WithChunkReading $import, Reader $reader, TemporaryFile $temporaryFile)
     {
-        if ($import instanceof WithEvents && isset($import->registerEvents()[BeforeImport::class])) {
+        if ($import instanceof WithEvents) {
             $reader->beforeImport($import);
         }
 
@@ -36,7 +49,7 @@ class ChunkReader
         $totalRows    = $reader->getTotalRows();
         $worksheets   = $reader->getWorksheets($import);
         $queue        = property_exists($import, 'queue') ? $import->queue : null;
-        $delayCleanup = property_exists($import, 'delayCleanup') ? $import->delayCleanup : 600;
+        $delayCleanup = property_exists($import, 'cleanupInterval') ? $import->cleanupInterval : 60;
 
         if ($import instanceof WithProgressBar) {
             $import->getConsoleOutput()->progressStart(array_sum($totalRows));
@@ -70,6 +83,8 @@ class ChunkReader
         $afterImportJob = new AfterImportJob($import, $reader);
 
         if ($import instanceof ShouldQueueWithoutChain) {
+            $afterImportJob->setInterval($delayCleanup);
+            $afterImportJob->setDependencies($jobs);
             $jobs->push($afterImportJob->delay($delayCleanup));
 
             return $jobs->each(function ($job) use ($queue) {
@@ -87,7 +102,9 @@ class ChunkReader
 
         $jobs->each(function ($job) {
             try {
-                dispatch_now($job);
+                function_exists('dispatch_now')
+                    ? dispatch_now($job)
+                    : $this->dispatchNow($job);
             } catch (Throwable $e) {
                 if (method_exists($job, 'failed')) {
                     $job->failed($e);
@@ -103,5 +120,27 @@ class ChunkReader
         unset($jobs);
 
         return null;
+    }
+
+    /**
+     * Dispatch a command to its appropriate handler in the current process without using the synchronous queue.
+     *
+     * @param  object  $command
+     * @param  mixed  $handler
+     * @return mixed
+     */
+    protected function dispatchNow($command, $handler = null)
+    {
+        $uses = class_uses_recursive($command);
+
+        if (in_array(InteractsWithQueue::class, $uses) &&
+            in_array(Queueable::class, $uses) && !$command->job
+        ) {
+            $command->setJob(new SyncJob($this->container, json_encode([]), 'sync', 'sync'));
+        }
+
+        $method = method_exists($command, 'handle') ? 'handle' : '__invoke';
+
+        return $this->container->call([$command, $method]);
     }
 }