+++ src/Imports/ModelImporter.php
@@ -7,15 +7,20 @@ use Maatwebsite\Excel\Concerns\ToModel;
 use Maatwebsite\Excel\Concerns\WithBatchInserts;
 use Maatwebsite\Excel\Concerns\WithCalculatedFormulas;
 use Maatwebsite\Excel\Concerns\WithColumnLimit;
+use Maatwebsite\Excel\Concerns\WithEvents;
 use Maatwebsite\Excel\Concerns\WithFormatData;
 use Maatwebsite\Excel\Concerns\WithMapping;
 use Maatwebsite\Excel\Concerns\WithProgressBar;
 use Maatwebsite\Excel\Concerns\WithValidation;
+use Maatwebsite\Excel\Events\AfterBatch;
+use Maatwebsite\Excel\HasEventBus;
 use Maatwebsite\Excel\Row;
 use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 
 class ModelImporter
 {
+    use HasEventBus;
+
     /**
      * @var ModelManager
      */
@@ -42,6 +47,9 @@ class ModelImporter
         if ($startRow > $worksheet->getHighestRow()) {
             return;
         }
+        if ($import instanceof WithEvents) {
+            $this->registerListeners($import->registerEvents());
+        }
 
         $headingRow       = HeadingRowExtractor::extract($worksheet, $import);
         $headerIsGrouped  = HeadingRowExtractor::extractGrouping($headingRow, $import);
@@ -56,13 +64,18 @@ class ModelImporter
 
         $this->manager->setRemembersRowNumber(method_exists($import, 'rememberRowNumber'));
 
-        $i = 0;
+        $i             = 0;
+        $batchStartRow = $startRow;
         foreach ($worksheet->getRowIterator($startRow, $endRow) as $spreadSheetRow) {
             $i++;
 
             $row = new Row($spreadSheetRow, $headingRow, $headerIsGrouped);
-            if (!$import instanceof SkipsEmptyRows || ($import instanceof SkipsEmptyRows && !$row->isEmpty($withCalcFormulas))) {
+            if (!$import instanceof SkipsEmptyRows || !$row->isEmpty($withCalcFormulas)) {
                 $rowArray = $row->toArray(null, $withCalcFormulas, $formatData, $endColumn);
+
+                if ($import instanceof SkipsEmptyRows && method_exists($import, 'isEmptyWhen') && $import->isEmptyWhen($rowArray)) {
+                    continue;
+                }
 
                 if ($withValidation) {
                     $rowArray = $import->prepareForValidation($rowArray, $row->getIndex());
@@ -79,7 +92,8 @@ class ModelImporter
 
                 // Flush each batch.
                 if (($i % $batchSize) === 0) {
-                    $this->manager->flush($import, $batchSize > 1);
+                    $this->flush($import, $batchSize, $batchStartRow);
+                    $batchStartRow += $i;
                     $i = 0;
 
                     if ($progessBar) {
@@ -89,7 +103,15 @@ class ModelImporter
             }
         }
 
-        // Flush left-overs.
+        if ($i > 0) {
+            // Flush left-overs.
+            $this->flush($import, $batchSize, $batchStartRow);
+        }
+    }
+
+    private function flush(ToModel $import, int $batchSize, int $startRow)
+    {
         $this->manager->flush($import, $batchSize > 1);
+        $this->raise(new AfterBatch($this->manager, $import, $batchSize, $startRow));
     }
 }