+++ src/Jobs/ReadChunk.php
@@ -5,9 +5,11 @@ namespace Maatwebsite\Excel\Jobs;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Support\Facades\Cache;
 use Maatwebsite\Excel\Concerns\WithChunkReading;
 use Maatwebsite\Excel\Concerns\WithCustomValueBinder;
 use Maatwebsite\Excel\Concerns\WithEvents;
+use Maatwebsite\Excel\Events\AfterChunk;
 use Maatwebsite\Excel\Events\ImportFailed;
 use Maatwebsite\Excel\Files\RemoteTemporaryFile;
 use Maatwebsite\Excel\Files\TemporaryFile;
@@ -38,6 +40,21 @@ class ReadChunk implements ShouldQueue
      * @var int
      */
     public $maxExceptions;
+
+    /**
+     * @var int
+     */
+    public $backoff;
+
+    /**
+     * @var string
+     */
+    public $queue;
+
+    /**
+     * @var string
+     */
+    public $connection;
 
     /**
      * @var WithChunkReading
@@ -75,6 +92,11 @@ class ReadChunk implements ShouldQueue
     private $chunkSize;
 
     /**
+     * @var string
+     */
+    private $uniqueId;
+
+    /**
      * @param  WithChunkReading  $import
      * @param  IReader  $reader
      * @param  TemporaryFile  $temporaryFile
@@ -95,6 +117,24 @@ class ReadChunk implements ShouldQueue
         $this->timeout       = $import->timeout ?? null;
         $this->tries         = $import->tries ?? null;
         $this->maxExceptions = $import->maxExceptions ?? null;
+        $this->backoff       = method_exists($import, 'backoff') ? $import->backoff() : ($import->backoff ?? null);
+        $this->connection    = property_exists($import, 'connection') ? $import->connection : null;
+        $this->queue         = property_exists($import, 'queue') ? $import->queue : null;
+    }
+
+    public function getUniqueId(): string
+    {
+        if (!isset($this->uniqueId)) {
+            $this->uniqueId = uniqid();
+            Cache::set('laravel-excel/read-chunk/' . $this->uniqueId, true);
+        }
+
+        return $this->uniqueId;
+    }
+
+    public static function isComplete(string $id): bool
+    {
+        return !Cache::has('laravel-excel/read-chunk/' . $id);
     }
 
     /**
@@ -129,6 +169,10 @@ class ReadChunk implements ShouldQueue
             $this->import->setChunkOffset($this->startRow);
         }
 
+        if (method_exists($this->sheetImport, 'setChunkOffset')) {
+            $this->sheetImport->setChunkOffset($this->startRow);
+        }
+
         if ($this->sheetImport instanceof WithCustomValueBinder) {
             Cell::setValueBinder($this->sheetImport);
         }
@@ -143,8 +187,8 @@ class ReadChunk implements ShouldQueue
         );
 
         $this->reader->setReadFilter($filter);
-        $this->reader->setReadDataOnly(true);
-        $this->reader->setReadEmptyCells(false);
+        $this->reader->setReadDataOnly(config('excel.imports.read_only', true));
+        $this->reader->setReadEmptyCells(!config('excel.imports.ignore_empty', false));
 
         $spreadsheet = $this->reader->load(
             $this->temporaryFile->sync()->getLocalPath()
@@ -172,6 +216,8 @@ class ReadChunk implements ShouldQueue
             $sheet->disconnect();
 
             $this->cleanUpTempFile();
+
+            $sheet->raise(new AfterChunk($sheet, $this->import, $this->startRow));
         });
     }
 
@@ -180,9 +226,7 @@ class ReadChunk implements ShouldQueue
      */
     public function failed(Throwable $e)
     {
-        if ($this->temporaryFile instanceof RemoteTemporaryFile) {
-            $this->temporaryFile->deleteLocalCopy();
-        }
+        $this->cleanUpTempFile(true);
 
         if ($this->import instanceof WithEvents) {
             $this->registerListeners($this->import->registerEvents());
@@ -194,9 +238,13 @@ class ReadChunk implements ShouldQueue
         }
     }
 
-    private function cleanUpTempFile()
+    private function cleanUpTempFile(bool $force = false): bool
     {
-        if (!config('excel.temporary_files.force_resync_remote')) {
+        if (!empty($this->uniqueId)) {
+            Cache::delete('laravel-excel/read-chunk/' . $this->uniqueId);
+        }
+
+        if (!$force && !config('excel.temporary_files.force_resync_remote')) {
             return true;
         }