+++ src/Row.php
@@ -18,6 +18,11 @@ class Row implements ArrayAccess
     protected $headingRow = [];
 
     /**
+     * @var array
+     */
+    protected $headerIsGrouped = [];
+
+    /**
      * @var \Closure
      */
     protected $preparationCallback;
@@ -36,6 +41,11 @@ class Row implements ArrayAccess
      * @var bool|null
      */
     protected $rowCacheFormatData;
+
+    /**
+     * @var string|null
+     */
+    protected $rowCacheEndColumn;
 
     /**
      * @param  SpreadsheetRow  $row
@@ -78,7 +88,7 @@ class Row implements ArrayAccess
      */
     public function toArray($nullValue = null, $calculateFormulas = false, $formatData = true, ?string $endColumn = null)
     {
-        if (is_array($this->rowCache) && ($this->rowCacheFormatData === $formatData)) {
+        if (is_array($this->rowCache) && ($this->rowCacheFormatData === $formatData) && ($this->rowCacheEndColumn === $endColumn)) {
             return $this->rowCache;
         }
 
@@ -107,6 +117,7 @@ class Row implements ArrayAccess
 
         $this->rowCache           = $cells;
         $this->rowCacheFormatData = $formatData;
+        $this->rowCacheEndColumn  = $endColumn;
 
         return $cells;
     }
@@ -132,13 +143,13 @@ class Row implements ArrayAccess
     #[\ReturnTypeWillChange]
     public function offsetExists($offset)
     {
-        return isset(($this->toArray())[$offset]);
+        return isset($this->toArray()[$offset]);
     }
 
     #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
-        return ($this->toArray())[$offset];
+        return $this->toArray()[$offset];
     }
 
     #[\ReturnTypeWillChange]