+++ tests/Concerns/WithChunkReadingTest.php
@@ -2,6 +2,7 @@
 
 namespace Maatwebsite\Excel\Tests\Concerns;
 
+use DateTime;
 use Exception;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
@@ -11,6 +12,7 @@ use Maatwebsite\Excel\Concerns\ToModel;
 use Maatwebsite\Excel\Concerns\WithBatchInserts;
 use Maatwebsite\Excel\Concerns\WithChunkReading;
 use Maatwebsite\Excel\Concerns\WithEvents;
+use Maatwebsite\Excel\Concerns\WithFormatData;
 use Maatwebsite\Excel\Concerns\WithHeadingRow;
 use Maatwebsite\Excel\Concerns\WithMultipleSheets;
 use Maatwebsite\Excel\Events\AfterImport;
@@ -20,6 +22,7 @@ use Maatwebsite\Excel\Reader;
 use Maatwebsite\Excel\Tests\Data\Stubs\Database\Group;
 use Maatwebsite\Excel\Tests\Data\Stubs\Database\User;
 use Maatwebsite\Excel\Tests\TestCase;
+use PhpOffice\PhpSpreadsheet\Shared\Date;
 use PHPUnit\Framework\Assert;
 use Throwable;
 
@@ -283,7 +286,7 @@ class WithChunkReadingTest extends TestCase
      */
     public function can_import_to_array_in_chunks()
     {
-        $import = new class implements ToArray, WithChunkReading
+        $import = new class implements ToArray, WithChunkReading, WithFormatData
         {
             use Importable;
 
@@ -521,5 +524,79 @@ class WithChunkReadingTest extends TestCase
         }
 
         $this->assertTrue($import->failed, 'ImportFailed event was not called.');
+    }
+
+    /**
+     * @test
+     */
+    public function can_import_to_array_and_format_in_chunks()
+    {
+        config()->set('excel.imports.read_only', false);
+
+        $import = new class implements ToArray, WithChunkReading, WithFormatData
+        {
+            use Importable;
+
+            /**
+             * @param  array  $array
+             */
+            public function array(array $array)
+            {
+                Assert::assertCount(2, $array);
+                Assert::assertCount(1, $array[0]);
+                Assert::assertCount(1, $array[1]);
+                Assert::assertIsString($array[0][0]);
+                Assert::assertIsString($array[1][0]);
+                Assert::assertEquals('01/12/22', $array[0][0]);
+                Assert::assertEquals('2023-02-20', $array[1][0]);
+            }
+
+            /**
+             * @return int
+             */
+            public function chunkSize(): int
+            {
+                return 2;
+            }
+        };
+
+        $import->import('import-batches-with-date.xlsx');
+    }
+
+    /**
+     * @test
+     */
+    public function can_import_to_array_in_chunks_without_formatting()
+    {
+        config()->set('excel.imports.read_only', true);
+
+        $import = new class implements ToArray, WithChunkReading
+        {
+            use Importable;
+
+            /**
+             * @param  array  $array
+             */
+            public function array(array $array)
+            {
+                Assert::assertCount(2, $array);
+                Assert::assertCount(1, $array[0]);
+                Assert::assertCount(1, $array[1]);
+                Assert::assertIsInt($array[0][0]);
+                Assert::assertIsInt($array[1][0]);
+                Assert::assertEquals((int) Date::dateTimeToExcel(DateTime::createFromFormat('Y-m-d', '2022-12-01')->setTime(0, 0, 0, 0)), $array[0][0]);
+                Assert::assertEquals((int) Date::dateTimeToExcel(DateTime::createFromFormat('Y-m-d', '2023-02-20')->setTime(0, 0, 0, 0)), $array[1][0]);
+            }
+
+            /**
+             * @return int
+             */
+            public function chunkSize(): int
+            {
+                return 2;
+            }
+        };
+
+        $import->import('import-batches-with-date.xlsx');
     }
 }