+++ CHANGELOG.md
@@ -2,10 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+Please view https://github.com/SpartnerNL/Laravel-Excel/releases for the most recent changelog
+
+## [3.1.47] - 2023-02-16
+
+- Support Laravel 10
+
+### Fixed
+
+- Fix Bug Multiple drawings change the behavior of the startCell (#3865).
+- Allow configuring read data only in chunks as well.
+
+## [3.1.46] - 2023-01-27
+
+- Support Laravel 10
+
+## [3.1.45] - 2023-01-02
 
 ### Added
 - Add support for ignoring PHP auto_detect_line_endings INI directive
+
+### Fixed
+- Fix the PSR simple cache dependency mess to maintain backwards compatability and support 3.0 of the interface.
 
 ## [3.1.44] - 2022-10-14
 
@@ -259,7 +277,14 @@ All notable changes to this project will be documented in this file.
 - Raw() method now also available on Exportable.
 - Fix for breaking changes in PhpSpreadsheet with empty enclosures.
 
-[Unreleased]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.40...HEAD
+[Unreleased]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.47...HEAD
+[3.1.47]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.46...3.1.47
+[3.1.46]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.45...3.1.46
+[3.1.45]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.44...3.1.45
+[3.1.44]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.43...3.1.44
+[3.1.43]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.42...3.1.43
+[3.1.42]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.41...3.1.42
+[3.1.41]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.40...3.1.41
 [3.1.40]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.39...3.1.40
 [3.1.39]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.38...3.1.39
 [3.1.38]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.37...3.1.38